migrated all usages to the new  package (#369)
