+++ cmd/zrok/configUnset.go
@@ -2,7 +2,8 @@ package main
 
 import (
 	"fmt"
-	"github.com/openziti/zrok/environment/env_v0_3"
+	"github.com/openziti/zrok/environment"
+	"github.com/openziti/zrok/environment/env_core"
 	"github.com/openziti/zrok/tui"
 	"github.com/spf13/cobra"
 	"os"
@@ -30,33 +31,23 @@ func newConfigUnsetCommand() *configUnsetCommand {
 func (cmd *configUnsetCommand) run(_ *cobra.Command, args []string) {
 	configName := args[0]
 
-	zrd, err := env_v0_3.Load()
+	env, err := environment.LoadRoot()
 	if err != nil {
 		panic(err)
 	}
 
-	modified := false
 	switch configName {
 	case "apiEndpoint":
-		if zrd.Cfg != nil && zrd.Cfg.ApiEndpoint != "" {
-			zrd.Cfg.ApiEndpoint = ""
-			modified = true
+		if err := env.SetConfig(&env_core.Config{}); err != nil {
+			tui.Error("unable to save config", err)
+		}
+		fmt.Println("zrok configuration updated")
+		if env.IsEnabled() {
+			fmt.Printf("\n[%v]: because you have a %v-d environment, you won't see your config change until you run %v first!\n\n", tui.WarningLabel, tui.Code.Render("zrok enable"), tui.Code.Render("zrok disable"))
 		}
 
 	default:
 		fmt.Printf("unknown config name '%v'\n", configName)
 		os.Exit(1)
-	}
-
-	if modified {
-		if err := zrd.Save(); err != nil {
-			panic(err)
-		}
-		fmt.Println("zrok configuration updated")
-		if zrd.Env != nil && configName == "apiEndpoint" {
-			fmt.Printf("\n[%v]: because you have a %v-d environment, you won't see your config change until you run %v first!\n\n", tui.WarningLabel, tui.Code.Render("zrok enable"), tui.Code.Render("zrok disable"))
-		}
-	} else {
-		fmt.Println("zrok configuration not changed")
 	}
 }
