+++ environment/env_v0_3/dirs.go
@@ -1,6 +1,7 @@
-package env_v0_3x
+package env_v0_3
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 )
@@ -35,4 +36,20 @@ func environmentFile() (string, error) {
 		return "", err
 	}
 	return filepath.Join(zrd, "environment.json"), nil
+}
+
+func identitiesDir() (string, error) {
+	zrd, err := rootDir()
+	if err != nil {
+		return "", err
+	}
+	return filepath.Join(zrd, "identities"), nil
+}
+
+func identityFile(name string) (string, error) {
+	idd, err := identitiesDir()
+	if err != nil {
+		return "", err
+	}
+	return filepath.Join(idd, fmt.Sprintf("%v.json", name)), nil
 }
