+++ sdk/share.go
@@ -1,8 +1,6 @@
 package sdk
 
 import (
-	"github.com/openziti/zrok/environment/env_v0_3"
-	"github.com/openziti/zrok/rest_model_zrok"
 	"github.com/pkg/errors"
 )
 
@@ -26,17 +24,5 @@ func newPrivateShare(request *ShareRequest) (*Share, error) {
 }
 
 func newPublicShare(request *ShareRequest) (*Share, error) {
-	return nil, nil
-}
-
-func loadEnvironment(request *ShareRequest) (*env_v0_3.Root, error) {
-	env, err := env_v0_3.Load()
-	if err != nil {
-		return nil, errors.Wrap(err, "error loading environment")
-	}
-	return env, nil
-}
-
-func createShare(zrd *env_v0_3.Root, req *rest_model_zrok.ShareRequest) (*Share, error) {
 	return nil, nil
 }
